+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.1.0
+## 1.1.0 2023-10-22
 
 - **New:**
   - Add a working javascript prompt() function